Move BookmarksFragment and BookmarksSortByDialog to bookmarks package
Also extract BookmarksPage outside the fragment
